+++ src/main/java/com/lazerycode/jmeter/mojo/RunJMeterMojo.java
@@ -14,7 +14,7 @@ import org.apache.maven.plugins.annotations.Mojo;
 import java.io.File;
 
 /**
- * Goal that runs jmeter based on configuration defined in your pom.<br/>
+ * Goal that runs jmeter based on configuration defined in your pom.<br>
  * This goal runs within Lifecycle phase {@link LifecyclePhase#INTEGRATION_TEST}.
  */
 @Mojo(name = "jmeter", defaultPhase = LifecyclePhase.INTEGRATION_TEST)
@@ -56,7 +56,7 @@ public class RunJMeterMojo extends AbstractJMeterMojo {
                 .setTestFilesExcluded(testFilesExcluded)
                 .setRemoteServerConfiguration(remoteConfig)
                 .setSuppressJMeterOutput(suppressJMeterOutput)
-                .setBinDir(new File(testConfig.getCurrentTestConfiguration().getJmeterWorkingDirectoryPath()))
+                .setBinDir(testConfig.getCurrentTestConfiguration().getJmeterWorkingDirectoryPath())
                 .setJMeterProcessJVMSettings(jMeterProcessJVMSettings)
                 .setRuntimeJarName(testConfig.getCurrentTestConfiguration().getRuntimeJarName())
                 .setReportDirectory(reportDirectory)